+++ src/components/Welcome/sections/Motivation.tsx
@@ -1,0 +1,30 @@
+export default function Motivation() {
+  return (
+    <div className="mb-20 max-w-3xl">
+      <h3 className="-translate-x-8 -skew-x-6 text-white text-opacity-40">
+        Motivation
+      </h3>
+      <div className="border-l-2 border-white border-opacity-30 p-3 px-8">
+        <p className="mb-4 mr-auto">
+          While working on some different projects, I was tired of just using a
+          solid color or some sort of subtle gradient for my backgrounds. Or
+          going through the hassle of creating some sort of image.
+        </p>
+        <p className="mb-4 mr-auto">
+          At the same time I was reading up on a lot of CSS Specifications Color
+          Level 3 & 4, and I thought it would be fun to build a tool that used
+          some of those features.
+        </p>
+        <p className="mb-4 mr-auto">
+          Taking inspiration from tools like Photoshop, I wanted to leverage the
+          power of the browser <i>(Chrome atleast)</i>, CSS, and HTML to create
+          a simple and intuitive tool for designing backgrounds.
+        </p>
+        <p className="mb-4 mr-auto">
+          I wanted something better, easier to use, and designed with the web in
+          mind.
+        </p>
+      </div>
+    </div>
+  )
+}