+++ src/components/index.ts
@@ -5,3 +5,10 @@ export { default as Footer } from './Footer'
 export { default as Controls } from './Controls'
 export { default as Pending } from './Pending'
 export { default as ErrorPage } from './ErrorPage'
+export { default as VisibilityController } from './VisibilityController'
+export {
+  useCapabilities,
+  CapabilitiesContext,
+  CapabilitiesProvider,
+} from './Capabilities'
+export { default as WelcomePage } from './Welcome'